refactor: abstract IndexedDB logic into separate module
