+++ projects/mafia-frontend/src/Home.tsx
@@ -20,64 +20,9 @@ import NightStageDoctorCommit from './states/NightStageDoctorCommit'
 import NightStageMafiaCommit from './states/NightStageMafiaCommit'
 
 import { useQuery } from '@tanstack/react-query'
-import { Address } from 'algosdk'
 import Navbar from './components/NavBar'
-// Import IDB
-import { DBSchema, openDB } from 'idb'
 
-// Database setup
-const DB_NAME = 'mafiaGameDB'
-const STORE_NAME = 'playerStore'
-const DB_VERSION = 1
-
-export interface Acct {
-  addr: Address
-  keyData: ArrayBuffer
-}
-
-export interface IDBPlayer {
-  day_algo_address: Acct
-
-  night_algo_address: Acct
-  commitment: Uint8Array | undefined
-
-  blinder: Uint8Array | undefined
-  bls_private_key: Uint8Array
-  salt: Uint8Array
-  iv: Uint8Array
-
-  bls_public_key: Uint8Array
-
-  target: string | undefined
-}
-
-interface WalletDB extends DBSchema {
-  playerStore: {
-    key: string
-    value: IDBPlayer
-  }
-}
-
-const initDB = async () => {
-  return openDB<WalletDB>(DB_NAME, DB_VERSION, {
-    upgrade(db) {
-      if (!db.objectStoreNames.contains(STORE_NAME)) {
-        db.createObjectStore(STORE_NAME)
-      }
-    },
-  })
-}
-
-const savePlayerData = async (playerId: string, playerData: IDBPlayer) => {
-  const db = await initDB()
-  await db.put(STORE_NAME, playerData, playerId)
-}
-
-const getPlayerData = async (playerId: string) => {
-  const db = await initDB()
-  const result = await db.get(STORE_NAME, playerId)
-  return result ? result : null
-}
+import { createStorageKey, getPlayerData, savePlayerData } from './db/playerStore'
 
 const Home: React.FC = () => {
   const [openWalletModal, setOpenWalletModal] = useState<boolean>(false)
@@ -91,29 +36,27 @@ const Home: React.FC = () => {
     setOpenWalletModal(!openWalletModal)
   }
 
+  // Initialize/restore Player Object from IndexedDB
   useEffect(() => {
     const initializePlayer = async () => {
       if (activeAddress && appId !== BigInt(0)) {
-        const storageKey = `${activeAddress}-${appId}`
+        const storageKey = createStorageKey(activeAddress, appId)
 
         try {
           // Try to get player data from IndexedDB
           const storedPlayerData = await getPlayerData(storageKey)
           if (storedPlayerData) {
             try {
-              const playerData = storedPlayerData
-              const loadedPlayer = await Player.fromIDB(playerData, appId)
+              const loadedPlayer = await Player.fromIDB(storedPlayerData, appId)
               setPlayerObject(loadedPlayer)
-
               console.log('Found stored player data', storedPlayerData)
               return
             } catch (error) {
               console.error('Error restoring player state:', error)
             }
           } else {
-            console.log('dont hit this else clause or ELSE!!!')
+            console.log('No stored player data found, creating new player')
             const player = new Player(appId)
-
             setPlayerObject(player)
           }
         } catch (error) {
@@ -128,14 +71,14 @@ const Home: React.FC = () => {
     initializePlayer()
   }, [activeAddress, appId])
 
+  // Save Player Object to IndexedDB when it changes
   useEffect(() => {
     const savePlayer = async () => {
       if (activeAddress && appId !== BigInt(0) && playerObject) {
-        const storageKey = `${activeAddress}-${appId}`
+        const storageKey = createStorageKey(activeAddress, appId)
 
         try {
           const playerData = await playerObject.toIDB()
-
           await savePlayerData(storageKey, playerData)
         } catch (error) {
           console.error('Error saving player state to IndexedDB:', error)
